Hack: switch to Unity.InputSystem for F6 key
This is a quick and dirty hack that uses Unity.InputSystem instead of UnityEngine.Input. This does not support the configuration files.

Solves error:
```
System.InvalidOperationException: You are trying to read Input using the UnityEngine.Input class, but you have switched active Input handling to Input System package in Player Settings.
  at (wrapper managed-to-native) UnityEngine.Input.GetKeyDownInt(UnityEngine.KeyCode)
  at UnityEngine.Input.GetKeyDown (UnityEngine.KeyCode key) [0x00001] in <3dc317ad8dc044888cd8672cc2923a88>:0
  at BepInEx.Configuration.KeyboardShortcut.IsDown () [0x0000c] in <23c82cada6174aeea1f2659206cab34d>:0
  at ScriptEngine.ScriptEngine.Update () [0x00002] in ScriptEngine.cs:46
```

Ideally we should detect which class should be used and respect the
configuration file.